+++ src/Futhark/Internalise/Monomorphise.hs
@@ -13,8 +13,6 @@
 --   is a side effect of the monomorphisation algorithm, which uses
 --   the entry points as roots).
 --
--- * Turns implicit record fields into explicit record fields.
---
 -- * Rewrite BinOp nodes to Apply nodes.
 --
 -- * Replace all size expressions by constants or variables,
@@ -36,14 +34,12 @@ import Data.Foldable
 import Data.List (partition)
 import Data.List.NonEmpty qualified as NE
 import Data.Map.Strict qualified as M
-import Data.Maybe
 import Data.Sequence qualified as Seq
 import Data.Set qualified as S
 import Futhark.MonadFreshNames
 import Futhark.Util (nubOrd, topologicalSort)
 import Futhark.Util.Pretty
 import Language.Futhark
-import Language.Futhark.Semantic (TypeBinding (..))
 import Language.Futhark.Traversals
 import Language.Futhark.TypeChecker.Types
 
@@ -133,25 +129,19 @@ entryAssert (x : xs) body =
 -- to a representation of their corresponding function bindings.
 data Env = Env
   { envPolyBindings :: M.Map VName PolyBinding,
-    envTypeBindings :: M.Map VName TypeBinding,
     envScope :: S.Set VName,
     envGlobalScope :: S.Set VName,
     envParametrized :: ExpReplacements
   }
 
 instance Semigroup Env where
-  Env tb1 pb1 sc1 gs1 pr1 <> Env tb2 pb2 sc2 gs2 pr2 = Env (tb1 <> tb2) (pb1 <> pb2) (sc1 <> sc2) (gs1 <> gs2) (pr1 <> pr2)
+  Env pb1 sc1 gs1 pr1 <> Env pb2 sc2 gs2 pr2 = Env (pb1 <> pb2) (sc1 <> sc2) (gs1 <> gs2) (pr1 <> pr2)
 
 instance Monoid Env where
-  mempty = Env mempty mempty mempty mempty mempty
+  mempty = Env mempty mempty mempty mempty
 
 localEnv :: Env -> MonoM a -> MonoM a
 localEnv env = local (env <>)
-
-extendEnv :: VName -> PolyBinding -> MonoM a -> MonoM a
-extendEnv vn binding =
-  localEnv
-    mempty {envPolyBindings = M.singleton vn binding}
 
 isolateNormalisation :: MonoM a -> MonoM a
 isolateNormalisation m = do
@@ -387,46 +377,45 @@ replaceExp e =
     maybeNormalisedSize _ = Nothing
 
 transformFName :: SrcLoc -> QualName VName -> StructType -> MonoM Exp
-transformFName loc fname t = do
-  t' <- removeTypeVariablesInType t
-  t'' <- transformType t'
-  let mono_t = monoType t'
+transformFName loc fname ft = do
+  t' <- transformType ft
+  let mono_t = monoType ft
   if baseTag (qualLeaf fname) <= maxIntrinsicTag
-    then pure $ var fname t''
+    then pure $ var fname t'
     else do
       maybe_fname <- lookupLifted (qualLeaf fname) mono_t
       maybe_funbind <- lookupFun $ qualLeaf fname
       case (maybe_fname, maybe_funbind) of
         -- The function has already been monomorphised.
         (Just (fname', infer), _) ->
-          applySizeArgs fname' (toRes Nonunique t'') <$> infer t''
+          applySizeArgs fname' (toRes Nonunique t') <$> infer t'
         -- An intrinsic function.
-        (Nothing, Nothing) -> pure $ var fname t''
+        (Nothing, Nothing) -> pure $ var fname t'
         -- A polymorphic function.
         (Nothing, Just funbind) -> do
           (fname', infer, funbind') <- monomorphiseBinding False funbind mono_t
           tell $ Seq.singleton (qualLeaf fname, funbind')
           addLifted (qualLeaf fname) mono_t (fname', infer)
-          applySizeArgs fname' (toRes Nonunique t'') <$> infer t''
+          applySizeArgs fname' (toRes Nonunique t') <$> infer t'
   where
-    var fname' t'' = Var fname' (Info t'') loc
+    var fname' t' = Var fname' (Info t') loc
 
-    applySizeArg t' (i, f) size_arg =
+    applySizeArg t (i, f) size_arg =
       ( i - 1,
         mkApply
           f
           [(Nothing, mempty, size_arg)]
-          (AppRes (foldFunType (replicate i i64) (RetType [] t')) [])
+          (AppRes (foldFunType (replicate i i64) (RetType [] t)) [])
       )
 
-    applySizeArgs fname' t' size_args =
+    applySizeArgs fname' t size_args =
       snd $
         foldl'
-          (applySizeArg t')
+          (applySizeArg t)
           ( length size_args - 1,
             Var
               (qualName fname')
-              (Info (foldFunType (map (const i64) size_args) (RetType [] t')))
+              (Info (foldFunType (map (const i64) size_args) (RetType [] t)))
               loc
           )
           size_args
@@ -507,30 +496,8 @@ transformAppExp (LetPat sizes pat e body loc) res = do
   body' <- withParams params $ scoping argset $ transformExp body
   res' <- transformAppRes res
   pure $ AppExp (LetPat sizes' pat' e' body' loc) (Info res')
-transformAppExp (LetFun fname (tparams, params, retdecl, Info ret, body) e loc) res
-  | not $ null tparams = do
-      -- Retrieve the lifted monomorphic function bindings that are
-      -- produced, filter those that are monomorphic versions of the
-      -- current let-bound function and insert them at this point, and
-      -- propagate the rest.
-      let funbind = PolyBinding (fname, tparams, params, ret, body, mempty, loc)
-      pass $ do
-        (e', bs) <- listen $ extendEnv fname funbind $ scoping (S.singleton fname) $ transformExp e
-        -- Do not remember this one for next time we monomorphise this
-        -- function.
-        modifyLifts $ filter ((/= fname) . fst . fst)
-        let (bs_local, bs_prop) = Seq.partition ((== fname) . fst) bs
-        pure (unfoldLetFuns (map snd $ toList bs_local) e', const bs_prop)
-  | otherwise = do
-      params' <- mapM transformPat params
-      body' <- scoping (S.fromList (foldMap patNames params')) $ transformExp body
-      ret' <- transformRetTypeSizes (S.fromList (foldMap patNames params')) ret
-      AppExp
-        <$> ( LetFun fname (tparams, params', retdecl, Info ret', body')
-                <$> scoping (S.singleton fname) (transformExp e)
-                <*> pure loc
-            )
-        <*> (Info <$> transformAppRes res)
+transformAppExp LetFun {} _ =
+  error "transformAppExp: LetFun is not supposed to occur"
 transformAppExp (If e1 e2 e3 loc) res =
   AppExp <$> (If <$> transformExp e1 <*> transformExp e2 <*> transformExp e3 <*> pure loc) <*> (Info <$> transformAppRes res)
 transformAppExp (Apply fe args _) res =
@@ -614,17 +581,8 @@ transformAppExp (BinOp (fname, _) (Info t) (e1, d1) (e2, d2) loc) res = do
         ( Var (qualName x) (Info argtype) mempty,
           Id x (Info argtype) mempty
         )
-transformAppExp (LetWith id1 id2 idxs e1 body loc) res = do
-  id1' <- transformIdent id1
-  id2' <- transformIdent id2
-  idxs' <- mapM transformDimIndex idxs
-  e1' <- transformExp e1
-  body' <- scoping (S.singleton $ identName id1') $ transformExp body
-  res' <- transformAppRes res
-  pure $ AppExp (LetWith id1' id2' idxs' e1' body' loc) (Info res')
-  where
-    transformIdent (Ident v t vloc) =
-      Ident v <$> traverse transformType t <*> pure vloc
+transformAppExp LetWith {} _ =
+  error "transformAppExp: LetWith is not supposed to occur"
 transformAppExp (Index e0 idxs loc) res =
   AppExp
     <$> (Index <$> transformExp e0 <*> mapM transformDimIndex idxs <*> pure loc)
@@ -862,16 +820,6 @@ desugarIndexSection idxs (Scalar (Arrow _ _ _ t1 (RetType dims t2))) loc = do
       loc
 desugarIndexSection _ t _ = error $ "desugarIndexSection: not a function type: " ++ prettyString t
 
--- Convert a collection of 'ValBind's to a nested sequence of let-bound,
--- monomorphic functions with the given expression at the bottom.
-unfoldLetFuns :: [ValBind] -> Exp -> Exp
-unfoldLetFuns [] e = e
-unfoldLetFuns (ValBind _ fname _ (Info rettype) dim_params params body _ _ loc : rest) e =
-  AppExp (LetFun fname (dim_params, params, Nothing, Info rettype, body) e' loc) (Info $ AppRes e_t mempty)
-  where
-    e' = unfoldLetFuns rest e
-    e_t = typeOf e'
-
 transformPat :: Pat (TypeBase Size u) -> MonoM (Pat (TypeBase Size u))
 transformPat = traverse transformType
 
@@ -1061,7 +1009,7 @@ monomorphiseBinding entry (PolyBinding (name, tparams, params, rettype, body, at
         substs' = M.map (Subst []) substs
         substStructType =
           substTypesAny (fmap (fmap (second (const mempty))) . (`M.lookup` substs'))
-        params' = map (substPat entry substStructType) params
+        params' = map (substPat substStructType) params
     params'' <- withArgs shape_names $ mapM transformPat params'
     exp_naming <- paramGetClean shape_names
 
@@ -1212,19 +1160,17 @@ typeSubstsM loc orig_t1 orig_t2 =
     onDim MonoAnon = pure anySize
 
 -- Perform a given substitution on the types in a pattern.
-substPat :: Bool -> (t -> t) -> Pat t -> Pat t
-substPat entry f pat = case pat of
-  TuplePat pats loc -> TuplePat (map (substPat entry f) pats) loc
+substPat :: (t -> t) -> Pat t -> Pat t
+substPat f pat = case pat of
+  TuplePat pats loc -> TuplePat (map (substPat f) pats) loc
   RecordPat fs loc -> RecordPat (map substField fs) loc
     where
-      substField (n, p) = (n, substPat entry f p)
-  PatParens p loc -> PatParens (substPat entry f p) loc
-  PatAttr attr p loc -> PatAttr attr (substPat entry f p) loc
+      substField (n, p) = (n, substPat f p)
+  PatParens p loc -> PatParens (substPat f p) loc
+  PatAttr attr p loc -> PatAttr attr (substPat f p) loc
   Id vn (Info tp) loc -> Id vn (Info $ f tp) loc
   Wildcard (Info tp) loc -> Wildcard (Info $ f tp) loc
-  PatAscription p td loc
-    | entry -> PatAscription (substPat False f p) td loc
-    | otherwise -> substPat False f p
+  PatAscription p _ _ -> substPat f p
   PatLit e (Info tp) loc -> PatLit e (Info $ f tp) loc
   PatConstr n (Info tp) ps loc -> PatConstr n (Info $ f tp) ps loc
 
@@ -1232,62 +1178,19 @@ toPolyBinding :: ValBind -> PolyBinding
 toPolyBinding (ValBind _ name _ (Info rettype) tparams params body _ attrs loc) =
   PolyBinding (name, tparams, params, rettype, body, attrs, loc)
 
--- Remove all type variables and type abbreviations from a value binding.
-removeTypeVariables :: Bool -> ValBind -> MonoM ValBind
-removeTypeVariables entry valbind = do
-  let (ValBind _ _ _ (Info (RetType dims rettype)) _ pats body _ _ _) = valbind
-  subs <- asks $ M.map substFromAbbr . envTypeBindings
-  let mapper =
-        ASTMapper
-          { mapOnExp = onExp,
-            mapOnName = pure,
-            mapOnStructType = pure . applySubst (`M.lookup` subs),
-            mapOnParamType = pure . applySubst (`M.lookup` subs),
-            mapOnResRetType = pure . applySubst (`M.lookup` subs)
-          }
-
-      onExp = astMap mapper
-
-  body' <- onExp body
-
-  pure
-    valbind
-      { valBindRetType = Info (applySubst (`M.lookup` subs) $ RetType dims rettype),
-        valBindParams = map (substPat entry $ applySubst (`M.lookup` subs)) pats,
-        valBindBody = body'
-      }
-
-removeTypeVariablesInType :: StructType -> MonoM StructType
-removeTypeVariablesInType t = do
-  subs <- asks $ M.map substFromAbbr . envTypeBindings
-  pure $ applySubst (`M.lookup` subs) t
-
-transformEntryPoint :: EntryPoint -> MonoM EntryPoint
-transformEntryPoint (EntryPoint params ret) =
-  EntryPoint <$> mapM onEntryParam params <*> onEntryType ret
-  where
-    onEntryParam (EntryParam v t) =
-      EntryParam v <$> onEntryType t
-    onEntryType (EntryType t te) =
-      EntryType <$> removeTypeVariablesInType t <*> pure te
-
 transformValBind :: ValBind -> MonoM Env
 transformValBind valbind = do
-  valbind' <-
-    toPolyBinding
-      <$> removeTypeVariables (isJust (valBindEntryPoint valbind)) valbind
+  let valbind' = toPolyBinding valbind
 
   case valBindEntryPoint valbind of
     Nothing -> pure ()
-    Just (Info entry) -> do
-      t <-
-        removeTypeVariablesInType $
-          funType (valBindParams valbind) $
-            unInfo $
-              valBindRetType valbind
+    Just entry -> do
+      let t =
+            funType (valBindParams valbind) $
+              unInfo $
+                valBindRetType valbind
       (name, infer, valbind'') <- monomorphiseBinding True valbind' $ monoType t
-      entry' <- transformEntryPoint entry
-      tell $ Seq.singleton (name, valbind'' {valBindEntryPoint = Just $ Info entry'})
+      tell $ Seq.singleton (name, valbind'' {valBindEntryPoint = Just entry})
       addLifted (valBindName valbind) (monoType t) (name, infer)
 
   pure
@@ -1299,30 +1202,15 @@ transformValBind valbind = do
             else mempty
       }
 
-transformTypeBind :: TypeBind -> MonoM Env
-transformTypeBind (TypeBind name l tparams _ (Info (RetType dims t)) _ _) = do
-  subs <- asks $ M.map substFromAbbr . envTypeBindings
-  let tbinding = TypeAbbr l tparams $ RetType dims $ applySubst (`M.lookup` subs) t
-  pure mempty {envTypeBindings = M.singleton name tbinding}
-
-transformDecs :: [Dec] -> MonoM ()
-transformDecs [] = pure ()
-transformDecs (ValDec valbind : ds) = do
+transformValBinds :: [ValBind] -> MonoM ()
+transformValBinds [] = pure ()
+transformValBinds (valbind : ds) = do
   env <- transformValBind valbind
-  localEnv env $ transformDecs ds
-transformDecs (TypeDec typebind : ds) = do
-  env <- transformTypeBind typebind
-  localEnv env $ transformDecs ds
-transformDecs (dec : _) =
-  error $
-    "The monomorphization module expects a module-free "
-      ++ "input program, but received: "
-      ++ prettyString dec
+  localEnv env $ transformValBinds ds
 
--- | Monomorphise a list of top-level declarations. A module-free input program
--- is expected, so only value declarations and type declaration are accepted.
-transformProg :: (MonadFreshNames m) => [Dec] -> m [ValBind]
+-- | Monomorphise a list of top-level value bindings.
+transformProg :: (MonadFreshNames m) => [ValBind] -> m [ValBind]
 transformProg decs =
   fmap (toList . fmap snd . snd) $
     modifyNameSource $ \namesrc ->
-      runMonoM namesrc $ transformDecs decs
+      runMonoM namesrc $ transformValBinds decs